Something that does not work at all
Originally this was intended to allow servers to define server-side-only argument type.
However, the implementation required the packet serialization format to be the same;
which caused problems, as the most convenient StringArgumentType is not usable in most
cases. Sync fallback proposal, RIP 2022-2022.

Note to self: never, ever, mess with registry sync.

Note to other modloader devs: if you can finish the implementation, I'd love to see one
(and integrate to Fabric API!) This is not tested at all, probably doesn't work, and the
main API isn't exposed yet. If you can figure out what this does, try it perhaps?